Force left-to-right rendering
When Vim tells us to draw a string with PUT, it expects us to put each
character to the right of the previous one, irrespective of its Unicode
directionality.

Before this patch, Neovim.app would sometimes appear to render RTL
text correctly when embedded in an LTR document, but this was not
consistent and was a bug.

It may be possible to implement true bidi support in the future, but
for now directionality is per-window as in Neovim.