Add MacSetFont function
This is the first NeoVim.app-specific functionality to be exposed to
VimL; implementation is with an RPC notification handled in window.mm.

When more of these start appearing, they should probably move out of
window.mm

Error reporting is via vim_report_error.

Also fixes save & restore of window size, which was broken with the
multiwindow patch. Only the size of the most recently-resized window
is saved.