Significantly improved redraw loop
The redraw loop was - as mentioned in a multi-line
comment - not efficient and the recent refactors and
optimisations made it possible to rewrite it
completely.

The background is now cleared in one step instead of
separately for each character and characters are
now not stored inside a vector-array but drawn
immediately.