Use rpc set_title event to set window title
This adds an event to the redraw action handler, that
sets the title appropriately.
The event was added to the gperf file.
The nvimrc was edited to output the event and the
correct title.

The method changing the window title was moved from
the window.mm file to the redraw.mm file and inlined.

The view.mm file now keeps a reference to the
NSWindow named mWindow.