Allow gnvim script to open relative paths
Apps opened with "-open" do not inherit the terminal's working
directory. Not a problem if you pass filenames to open directly to
"open", e.g. "open -a Neovim somefile". But, we want the full power
of Neovim's CLI arguments so we need to use e.g.
"open -a Neovim --args +287 somefile". This means open does not
intervene and interpret paths relative to $PWD.

Solution: Add a new CLI arg --cwd which tells Neovim.app to set
its working dir before opening files.

It's not necessarily desirable to have this working dir changed just
because you wanted to open some files from the terminal. Maybe we
should change the working dir back after opening the files? Not sure.

Leaving it like this for now.