Start with an empty environment when setting the shell environment
The neovim-dot-app process inherits whatever environment variables are
present when it is started. This leads to different behavior when it
is started from the command line or by clicking on the app icon.

Invoking env with -i erases the current environment, before starting a
shell in login mode to probe for the shell's environment. Fixes #84.