Add preference to open files in new tabs
Fixes #182. Adds a "MacSetOpenFilePreference" vim function, which sends a
"neovim.app.setOpenFilePreference" message to the GUI. If the argument to the
function is "new-tab", files dragged to the app/used via "open with" or gnvim
will open in new tabs; if it's "new window", a new window will open and the
files will be opened in tabs in that window.