use pkg-config to generate CFLAGS and LDFLAGS
Rather than requiring the user to set CFLAGS and LDFLAGS in the event
that msgpack is not installed down `/usr/local` (this happens _e.g._
when someone has a nonstandard Homebrew installation, see #184 for an
example), use SCons' native support for `pkg-config` to set those
variables automatically.

If `pkg-config` is unavailable (perhaps the user has compiled msgpack by
hand), SCons will trap the error and proceed, and the user must set the
build variables to support their installation.