Switch to Neovim's new D- modifier for Command key
Up until now, we've been using <T-> (“Meta”) for the Command key. This
works fine, but Neovim has now added a dedicated Command modifier, <D->.

This patch makes the following changes:
- Command key now sends <D-> to Neovim, and not <T->.
- Refuse to build using versions of Neovim before this change (or else
  we'd have to support both modifiers.)
- On startup, show an error if user init file has mapped <T->.
- MacMenu accepts both, because that was easy to do.

It was not feasible to send both <T-> and <D-> for a single key-press
because this would break <C-o> in insert mode, and possibly other
things.