Make gnvim suck less
The idea here is to detect if the arguments to gnvim include options
or if they're just filenames. If filenames, pass them directly to
"open", which will take care of sending them to a running Neovim
instance if present. If options are included, start a new instance.

Not perfect but better than before.