Remove Homebrew instructions from README
Installing this way doesn't currently work, since our Homebrew formula depends
on the old Python 2 SCons, but Homebrew have since swapped out the
'scons' formula for the non-backwards-compatible Python 3 SCons.

Homebrew is a bit of a pain with its lack of dependency versioning, to be
honest. Might be time to reconsider supporting it.